pyproject.toml
[project]
name = "auth-fast-api-template"
version = "1.0.0"
description = "Production-ready FastAPI authentication system with JWT, OAuth2, and PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["fastapi", "authentication", "jwt", "oauth2", "postgresql"]

dependencies = [
    # FastAPI & Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.0",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "authlib>=1.4.0",
    "httpx>=0.28.0",
    
    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    
    # Rate Limiting (in-memory)
    "slowapi>=0.1.9",
    
    # Utilities
    "python-multipart>=0.0.18",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
]
ignore = ["S101", "S104"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
